AnagramTests
Some renaming refactoring and created AnagramTests which is JUnit test
to validate that output when input is "holyoke" is as expected. (Output
seems to be displaying funny for me, with a lot of line breaks, do you
see this as well? There weren't so many line breaks when I originally
ran the project files on my computer.)